+++ src/main/java/com/jaehong/projectclassjaehongdev/post/service/PostDeleteServiceImpl.java
@@ -1,6 +1,6 @@
 package com.jaehong.projectclassjaehongdev.post.service;
 
-import com.jaehong.projectclassjaehongdev.config.domain.DomainExceptionCode;
+import com.jaehong.projectclassjaehongdev.global.domain.DomainExceptionCode;
 import com.jaehong.projectclassjaehongdev.post.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
